requirements.txt
#BE SURE TO ADD ANY LIBRARIES YOU USE TO THE REQUIREMENTS.TXT FILE!

streamlit
geopandas
pandas
matplotlib
pandas
folium
streamlit_foliumstreamlit 
numpy
contextily
zipfile
os
glob
graphviz
warnings
seaborn
plotly.express
io
gmplot
simplekml
keyring
yagmail